+++ src/emu/emu_ev.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021-2023 Barcelona Supercomputing Center (BSC)
+/* Copyright (c) 2021-2024 Barcelona Supercomputing Center (BSC)
  * SPDX-License-Identifier: GPL-3.0-or-later */
 
 #include "emu_ev.h"
@@ -13,11 +13,11 @@ emu_ev(struct emu_ev *ev, const struct ovni_ev *oev,
 	ev->v = oev->header.value;
 	ev->mcv[3] = '\0';
 
-	ev->rclock = oev->header.clock;
+	ev->rclock = (int64_t) oev->header.clock;
 	ev->sclock = sclock;
 	ev->dclock = dclock;
 
-	ev->payload_size = ovni_payload_size(oev);
+	ev->payload_size = (size_t) ovni_payload_size(oev);
 
 	if (ev->payload_size > 0) {
 		ev->has_payload = 1;
